requirements.txt
psycopg2-binary
pandas
minio
boto3
s3fs
jupyter
